// src/app/administrative/models/docente.ts
export interface Docente{
  id:number;
  identificacionDocente: number;
  tipoIdentificacionDocente: string;
  nombreDocente: string;
  apellidoDocente: string;
  correoElectronicoDocente: string;
  telefonoCelularDocente: string;
  numeroContratoDocente: string;
  ciudadResidenciaDocente: string;
  escalafonTecnicoDocente: string;
  escalafonExtensionDocente: string;
}